extract lookup and storeId to common
