+++ lib/app.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const common = require('./common');
-const BASE_URL = 'https://itunes.apple.com/lookup';
 
 function app (opts) {
   return new Promise(function (resolve) {
@@ -10,19 +9,14 @@ function app (opts) {
     }
     const idField = opts.id ? 'id' : 'bundleId';
     const idValue = opts.id || opts.appId;
-    const country = opts.country || 'us';
-    return resolve(`${BASE_URL}?${idField}=${idValue}&country=${country}`);
+    resolve(common.lookup([idValue], idField, opts.country));
   })
-  .then(common.request)
-  .then(JSON.parse)
-  .then((res) => {
-    return new Promise((resolve) => {
-      if (res.results.length === 0) {
-        throw Error('App not found (404)');
-      }
+  .then((results) => {
+    if (results.length === 0) {
+      throw Error('App not found (404)');
+    }
 
-      resolve(common.cleanApp(res.results[0]));
-    });
+    return results[0];
   });
 }
