+++ lib/common.js
@@ -2,6 +2,7 @@
 
 const request = require('request');
 const debug = require('debug')('app-store-scraper');
+const c = require('./constants');
 
 function cleanApp (app) {
   return {
@@ -58,4 +59,22 @@ const doRequest = (url, headers) => new Promise(function (resolve, reject) {
   });
 });
 
-module.exports = {cleanApp, request: doRequest};
+const LOOKUP_URL = 'https://itunes.apple.com/lookup';
+
+function lookup (ids, idField, country) {
+  idField = idField || 'id';
+  country = country || 'us';
+  const joinedIds = ids.join(',');
+  const url = `${LOOKUP_URL}?${idField}=${joinedIds}&country=${country}`;
+  return doRequest(url)
+    .then(JSON.parse)
+    .then((res) => res.results.map(cleanApp));
+}
+
+function storeId (countryCode) {
+  const markets = c.markets;
+  const defaultStore = '143441';
+  return (countryCode && markets[countryCode.toUpperCase()]) || defaultStore;
+}
+
+module.exports = {cleanApp, lookup, request: doRequest, storeId};
